requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy==1.21.0
pillow==8.3.1
